move inject_header_content into a module
