+++ lib/vbms.rb
@@ -26,6 +26,7 @@ require "vbms/responses/contention"
 require "vbms/responses/disposition"
 
 require "vbms/requests/base_request"
+require "vbms/requests/add_ext_security_header"
 # eDocument Service v4
 require "vbms/requests/upload_document_with_associations"
 require "vbms/requests/list_documents"
