+++ lib/vbms/requests/establish_claim.rb
@@ -1,6 +1,8 @@
 module VBMS
   module Requests
     class EstablishClaim < BaseRequest
+      include AddExtSecurityHeader
+
       NAMESPACES = {
         "xmlns:cla" => "http://vbms.vba.va.gov/external/ClaimService/v4",
         "xmlns:cdm" => "http://vbms.vba.va.gov/cdm/claim/v4",
@@ -30,19 +32,6 @@ module VBMS
 
       def endpoint_url(base_url)
         "#{base_url}#{VBMS::ENDPOINTS[specify_endpoint]}"
-      end
-
-      def inject_header_content(header_xml)
-        Nokogiri::XML::Builder.with(header_xml) do |xml|
-          xml["vbmsext"].userId("dslogon.1011239249", "xmlns:vbmsext" => "http://vbms.vba.va.gov/external")
-          if @send_userid
-            xml["ext"].Security("xmlns:ext" => "http://vbms.vba.va.gov/external") do
-              xml["ext"].cssUserName # do not insert css id in here, this will be injected by the client
-              xml["ext"].cssStationId
-              xml["ext"].SecurityLevel security_level
-            end
-          end
-        end
       end
 
       # More information on what the fields mean, see:
